Add silence spell spec
Uses the In-Code spell resistance designation I found. This parses to 2.
Right now it acts the same as `No`, but perhaps the system could be
revised to make some things more convenient.

I also correctly marked the targeting as `Area or Obj`. This parses as
Area right now, and there is hard coding to fix it back to `Area or
Obj`, but this is the actual correct designation.