Make magic missile interact with mirror image on strict rules
Missiles target images as well as the image caster, so targeting is
randomly determined as usual. It is assumed that the missile caster will
always target distinct copies, to avoid wasting missiles if possible.

This makes mirror image a defense against magic missile, but in turn
allows magic missile to quickly take out images.