Eliminate XP setting from Raise Dead script
The original temple dll set the XP, but it seems like the value it set
was incorrect. I've made the reimplementation set the right XP in the
case where it isn't adding a negative level (which already resets the
target's XP). This should also have the effect of making the set XP
value correct w/r/t level drain, level adjustment, and racial hit dice,
which the script wasn't (because I don't think it has any easy way to
get that number right now).