Replace Jaroo permanent spell implementation
The existing one used his Reincarnate spell to clear the character's
status by killing them first. Now that will add a negative level on
strict rules. That could be worked around, but instead I just changed
the script to fake him casting the spell, and used the Cuthbert
resurrection mode which has no negative effects and is completely
unrestricted. This makes it insensitive to any changes to Reincarnate,
as well.