Rework encumbrance dex AC capping
Previously they were interacting with the MaxDexAC dispatch for armor,
so being burdened could make the armor look like it has a lower max dex
value (unless a different query was used).

This replaces those callbacks to instead respond to the AC query, and
install their cap directly. This should make it clear that the
encumbrance is the source of the cap (not the armor).

Also, the caps were wrong for some encumbrance conditions, so those have
been fixed.