Add armor check adjustment for enhancement conditions
Armor and shield enhancements mark items as masterwork, because crafting
strips off the masterwork condition. The armor check reduction is
guarded by checking for a spell id, though, because the same conditions
are used for temporary buff spells which should not make the armor
masterwork.

Switched the bonus type of the adjustment to 12, 'enhancement bonus,' to
prevent stacking in case an item ends up with multiple
masterwork/enhancement conditions.