Rework attack sequence related to loading weapons
The previous code was dropping the projectile animation for
crossbows when fired during a full attack. The projectile
animation is what causes the bow to become unloaded, so a full
attack would result in all attacks even for heavy crossbows,
because it would never become unloaded.

The problem was that the ranged attack flag was being set on the
wrong d20 action in this code path. But I've decided to keep the
reimplementation of some assembly wrapper stuff.

This also adds infrastructure for slings requiring loading under
strict rules.