Fix two weapon animation logic
When checking whether a character has something in their off hand,
the shield logic was overwriting the weapon logic. This is a problem
with a buckler, which can be worn with an off hand weapon.

The result was using the two-handed animation for one-handed weapons
when wielding an off hand weapon with a buckler. The actual attack/
damage logic seems to have been unaffected, because there is
additional logic on top. So only animations were affected.