Fix a crash in natural_bond.py
math.floor was being called on the calculated animal companion level.
All this actually seems to do is convert the number to a float, because
the original calculation is ranger_level/2, which is integer division,
and automatically truncates.

Later, this value is used with += on an integer-backed field from C++,
which crashes with a type error (only int supported). This was causing
strange errors in seemingly unrelated situations for reasons I don't
completely understand.