Fix helpless/paralyze stat cap
The original function seemed to be implementing the wrong thing. It
capped _bonuses_ by 0, but allowed penalties. This meant you could end
up with negative stats, because the base is treated as a bonus, but
penalties are not cleared. The correct thing seems to be to set an
overall cap of 0.