Import tweaked version of Color Spray script
This version removes the 1d6 target cap (holdover from 3.0) and changes
the print statements to calls to a new, simple logging module for spell
scripts.

The log module checks the Temple+ log level to determine whether to
actually print, so most of the spammy messages in spell scripts could
eventually be replaced with `trace()` calls to avoid all the noise.