Make racial conditions pull ability modifiers from race data
The race data is defined via python, so is more easily modded. However,
the core race conditions were hard coded for ability mods, so the python
data would only apply to character creation. This makes the python
settings affect the core races same as the new races (which
programmatically add hooks based on the python settings).