Divide XP before putting it in the ego pile (#911)
The logbook tracks the most XP gained in a single fight. The calculation
for this was adding the _full_ XP for each opponent for each character,
rather than just that character's share. This ends up multiplying the
recorded XP by the number of (surviving) party members, rather than
recording the actual XP gained.